requirements.txt
PyYAML==3.10
chardet==2.1.1
localizable==0.1.2
wsgiref==0.1.2
